+++ libcotrik/src/PhongProjector.h
@@ -17,13 +17,13 @@ public:
     virtual ~PhongProjector();
 
 public:
-    void ProjectPointOnQuad(const Face& quad, const glm::vec3& p, glm::vec2& uv, glm::vec3& interpolP, glm::vec3& interpolN);
-    void ProjectPointOnTriangle(const Face& tri, const glm::vec3& p, glm::vec2& uv, glm::vec3& interpolP, glm::vec3& interpolN);
+    void ProjectPointOnQuad(const Face& quad, const glm::dvec3& p, glm::dvec2& uv, glm::dvec3& interpolP, glm::dvec3& interpolN);
+    void ProjectPointOnTriangle(const Face& tri, const glm::dvec3& p, glm::dvec2& uv, glm::dvec3& interpolP, glm::dvec3& interpolN);
 private:
     const Mesh& mesh;
 };
 
-glm::vec3 bilinear(const glm::vec3& v1, const glm::vec3& v2, const glm::vec3& v3, const glm::vec3& v4, const glm::vec2& uv);
-glm::vec3 barycentric(const glm::vec3& v1, const glm::vec3& v2, const glm::vec3& v3, const glm::vec2& uv);
+glm::dvec3 bilinear(const glm::dvec3& v1, const glm::dvec3& v2, const glm::dvec3& v3, const glm::dvec3& v4, const glm::dvec2& uv);
+glm::dvec3 barycentric(const glm::dvec3& v1, const glm::dvec3& v2, const glm::dvec3& v3, const glm::dvec2& uv);
 
 #endif /* LIBCOTRIK_SRC_PHONGPROJECTOR_H_ */
